chore: move mempool_infra tests
