+++ crates/mempool_infra/src/lib.rs
@@ -3,4 +3,6 @@ pub mod component_definitions;
 pub mod component_server;
 pub mod errors;
 pub mod serde_utils;
+#[cfg(test)]
+pub mod tests;
 pub mod trace_util;
